Fixing "previousAttributes" bug;
We had a property on the model that toggled the showing of the loading
icon.  This property wasn't actually part of the model, so we were
setting it manually and triggering the 'change' event.  Unfortunately,
this didn't seem to work in some environments. So now we've added an
attribute that is part of the model and are setting it properly, which
triggers the correct events.